refactor(tabs): extract cycleTab as exported pure function
Test now imports the production function instead of duplicating logic.
Added zero-tabs edge case.
