+++ src/client/hooks/useTabCycleKeyboard.ts
@@ -1,5 +1,18 @@
 import { useEffect, useRef } from "react";
 
+/** Pure cycling logic — exported for direct testing. */
+export function cycleTab(
+  tabs: Array<{ id: string }>,
+  activeTabId: string | null,
+  shiftKey: boolean,
+): string | null {
+  if (tabs.length < 2) return null;
+  const currentIdx = tabs.findIndex((t) => t.id === activeTabId);
+  const direction = shiftKey ? -1 : 1;
+  const nextIdx = (currentIdx + direction + tabs.length) % tabs.length;
+  return tabs[nextIdx].id;
+}
+
 /**
  * Ctrl+Tab / Ctrl+Shift+Tab to cycle through document tabs.
  * Registers a single listener on mount; refs avoid churn on every tab switch.
@@ -18,15 +31,12 @@ export function useTabCycleKeyboard(
     const handler = (e: KeyboardEvent) => {
       if (!e.ctrlKey && !e.metaKey) return;
       if (e.key !== "Tab") return;
-      const tabs = tabsRef.current;
-      if (tabs.length < 2) return;
+
+      const nextId = cycleTab(tabsRef.current, activeRef.current, e.shiftKey);
+      if (!nextId) return;
 
       e.preventDefault();
-
-      const currentIdx = tabs.findIndex((t) => t.id === activeRef.current);
-      const direction = e.shiftKey ? -1 : 1;
-      const nextIdx = (currentIdx + direction + tabs.length) % tabs.length;
-      setActiveTabId(tabs[nextIdx].id);
+      setActiveTabId(nextId);
     };
     window.addEventListener("keydown", handler);
     return () => window.removeEventListener("keydown", handler);
